Refactor: extract dev link shortcuts from homepage into its own component
